// src/modules/auth/types/types.messages.ts
export const enum typesMessages {
  LOGIN = `Te has logueado satisfactoriamente.`,
  INVALIDCREDENTIALS = `Los datos introducidos no son validos!`,
  DUPLICATE = 'El usuario ya esta esta registrado.',
  USER = 'El usuario',
  EVENT = 'El evento',
  REGISTER = `ha sido registrado satisfactoriamente.`,
  EVENTCREATED = 'El evento ha sido creado satisfactoriamente.',
  DELETED = 'ha sido borrado satisfactoriamente.',
  NOTFOUND = 'no ha podido ser encontrado.',
  UPDATED = 'ha sido actualizado correctamente.',
  JOINED = 'ha sido unido satisfactoriamente',
  COMMENT = 'El comentario',
  CREATED = 'ha sido creado satisfactoriamente',
  TODO = 'El todo',
}